refactor(ui): chip linkchip avatarchip linkavatarchip
